+++ weblate/auth/permissions.py
@@ -209,6 +209,8 @@ def check_contribute(user, permission, translation):
 def check_machinery(user, permission, obj):
     if not MACHINE_TRANSLATION_SERVICES.exists():
         return False
+    if isinstance(obj, Translation) and obj.is_template:
+        return False
     return check_contribute(user, permission, obj)