+++ weblate/formats/ttkit.py
@@ -21,7 +21,6 @@
 
 from __future__ import unicode_literals
 
-import csv
 import importlib
 import re
 
@@ -29,6 +28,7 @@ from django.utils.translation import ugettext_lazy as _
 
 import six
 
+from translate.storage.csvl10n import csv
 from translate.storage.po import pounit, pofile
 from translate.storage.poheader import default_header
 from translate.storage.ts2 import tsfile, tsunit
@@ -329,10 +329,11 @@ class XliffFormat(FileFormat):
 
     def find_matching(self, template_unit):
         """Find matching store unit for template"""
-        return self._find_unit_mono(
-            template_unit.source,
-            self.store
-        )
+        for search_unit in self.store.units:
+            if search_unit.source == template_unit.source:
+                return search_unit
+
+        return None
 
     def find_unit(self, context, source):
         return super(XliffFormat, self).find_unit(
@@ -344,6 +345,12 @@ class XliffFormat(FileFormat):
             self.unit_class(self._find_unit_mono(context, self.store)),
             False
         )
+
+    def create_unit(self, key, source):
+        unit = super(XliffFormat, self).create_unit(key, source)
+        unit.marktranslated()
+        unit.markapproved(False)
+        return unit
 
 
 class PoXliffFormat(XliffFormat):
@@ -561,7 +568,10 @@ class CSVFormat(FileFormat):
         if store.fieldnames != ['location', 'source', 'target']:
             return store
 
-        if not isinstance(content, six.string_types) and six.PY3:
+        # Do we have python 3 compatible csv module?
+        new_csv = six.PY3 or hasattr(csv, 'PY3')
+
+        if new_csv:
             content = content.decode('utf-8')
 
         fileobj = csv.StringIO(content)
@@ -577,7 +587,7 @@ class CSVFormat(FileFormat):
             return store
 
         result = storeclass(fieldnames=['source', 'target'])
-        if six.PY3:
+        if new_csv:
             result.parse(content.encode('utf-8'))
         else:
             result.parse(content)