+++ weblate/trans/tests/test_views.py
@@ -70,6 +70,7 @@ class RegistrationTestMixin(object):
                 return line[18:]
 
         self.fail('Confirmation URL not found')
+        return ''
 
     def assert_notify_mailbox(self, sent_mail):
         self.assertEqual(
@@ -208,7 +209,7 @@ class ViewTestCase(RepoTestCase):
     def assert_backend(self, expected_translated):
         """Check that backend has correct data."""
         translation = self.get_translation()
-        translation.commit_pending(None)
+        translation.commit_pending('test', None)
         store = translation.component.file_format_cls(
             translation.get_filename(),
             None